requirements.txt
biopython
pandas
numpy
nltk
spacy
scispacy
gensim
pyLDAvis
